+++ collector-bitmovin-player/src/main/java/com/bitmovin/analytics/bitmovin/player/BitmovinSdkAdapter.java
@@ -11,11 +11,13 @@ import com.bitmovin.analytics.adapters.PlayerAdapter;
 import com.bitmovin.analytics.BitmovinAnalyticsConfig;
 import com.bitmovin.analytics.data.ErrorCode;
 import com.bitmovin.analytics.data.EventData;
+import com.bitmovin.analytics.data.LanguageInformation;
 import com.bitmovin.analytics.enums.PlayerType;
 import com.bitmovin.analytics.stateMachines.PlayerState;
 import com.bitmovin.analytics.stateMachines.PlayerStateMachine;
 import com.bitmovin.analytics.utils.Util;
-import com.bitmovin.player.*;
+import com.bitmovin.player.BitmovinPlayer;
+import com.bitmovin.player.api.event.data.AudioChangedEvent;
 import com.bitmovin.player.api.event.data.AudioPlaybackQualityChangedEvent;
 import com.bitmovin.player.api.event.data.DroppedVideoFramesEvent;
 import com.bitmovin.player.api.event.data.ErrorEvent;
@@ -29,7 +31,9 @@ import com.bitmovin.player.api.event.data.SourceLoadedEvent;
 import com.bitmovin.player.api.event.data.SourceUnloadedEvent;
 import com.bitmovin.player.api.event.data.StallEndedEvent;
 import com.bitmovin.player.api.event.data.StallStartedEvent;
+import com.bitmovin.player.api.event.data.SubtitleChangedEvent;
 import com.bitmovin.player.api.event.data.VideoPlaybackQualityChangedEvent;
+import com.bitmovin.player.api.event.listener.OnAudioChangedListener;
 import com.bitmovin.player.api.event.listener.OnAudioPlaybackQualityChangedListener;
 import com.bitmovin.player.api.event.listener.OnDroppedVideoFramesListener;
 import com.bitmovin.player.api.event.listener.OnErrorListener;
@@ -43,10 +47,13 @@ import com.bitmovin.player.api.event.listener.OnSourceLoadedListener;
 import com.bitmovin.player.api.event.listener.OnSourceUnloadedListener;
 import com.bitmovin.player.api.event.listener.OnStallEndedListener;
 import com.bitmovin.player.api.event.listener.OnStallStartedListener;
+import com.bitmovin.player.api.event.listener.OnSubtitleChangedListener;
 import com.bitmovin.player.api.event.listener.OnVideoPlaybackQualityChangedListener;
 import com.bitmovin.player.config.media.SourceItem;
 import com.bitmovin.player.config.quality.AudioQuality;
 import com.bitmovin.player.config.quality.VideoQuality;
+import com.bitmovin.player.config.track.AudioTrack;
+import com.bitmovin.player.config.track.SubtitleTrack;
 
 public class BitmovinSdkAdapter implements PlayerAdapter {
     private static final String TAG = "BitmovinPlayerAdapter";
@@ -88,6 +95,8 @@ public class BitmovinSdkAdapter implements PlayerAdapter {
         this.bitmovinPlayer.addEventListener(onVideoPlaybackQualityChangedListener);
         this.bitmovinPlayer.addEventListener(onAudioPlaybackQualityChangedListener);
         this.bitmovinPlayer.addEventListener(onDroppedVideoFramesListener);
+        this.bitmovinPlayer.addEventListener(onSubtitleChangedListener);
+        this.bitmovinPlayer.addEventListener(onAudioChangedListener);
 
         this.bitmovinPlayer.addEventListener(onErrorListener);
     }
@@ -109,6 +118,8 @@ public class BitmovinSdkAdapter implements PlayerAdapter {
         this.bitmovinPlayer.removeEventListener(onAudioPlaybackQualityChangedListener);
         this.bitmovinPlayer.removeEventListener(onDroppedVideoFramesListener);
         this.bitmovinPlayer.removeEventListener(onErrorListener);
+        this.bitmovinPlayer.removeEventListener(onSubtitleChangedListener);
+        this.bitmovinPlayer.removeEventListener(onAudioChangedListener);
     }
 
     private String getUserAgent(Context context) {
@@ -207,6 +218,17 @@ public class BitmovinSdkAdapter implements PlayerAdapter {
             data.setAudioCodec(audioQuality.getCodec());
         }
 
+        //Subtitle info
+        SubtitleTrack subtitle = bitmovinPlayer.getSubtitle();
+        if (subtitle != null && subtitle.getId() != null) {
+            data.setSubtitleLanguage(subtitle.getLanguage());
+        }
+
+        //Audio language
+        AudioTrack audioTrack = bitmovinPlayer.getAudio();
+        if (audioTrack != null && audioTrack.getId() != null) {
+            data.setAudioLanguage(audioTrack.getLanguage());
+        }
 
         return data;
     }
@@ -295,6 +317,30 @@ public class BitmovinSdkAdapter implements PlayerAdapter {
         }
     };
 
+    private OnAudioChangedListener onAudioChangedListener = new OnAudioChangedListener() {
+        @Override
+        public void onAudioChanged(AudioChangedEvent audioChangedEvent) {
+            Log.d(TAG, "On AudioChanged: " + bitmovinPlayer.getAudio().getId());
+            if ((stateMachine.getCurrentState() == PlayerState.PLAYING) || (stateMachine.getCurrentState() == PlayerState.PAUSE)) {
+                PlayerState originalState = stateMachine.getCurrentState();
+                stateMachine.transitionState(PlayerState.AUDIOTRACKCHANGE, getPosition());
+                stateMachine.transitionState(originalState, getPosition());
+            }
+        }
+    };
+
+    private OnSubtitleChangedListener onSubtitleChangedListener = new OnSubtitleChangedListener() {
+        @Override
+        public void onSubtitleChanged(SubtitleChangedEvent event) {
+            Log.d(TAG, "On SubtitleChanged: " + bitmovinPlayer.getSubtitle().getId());
+            if ((stateMachine.getCurrentState() == PlayerState.PLAYING) || (stateMachine.getCurrentState() == PlayerState.PAUSE)) {
+                PlayerState originalState = stateMachine.getCurrentState();
+                stateMachine.transitionState(PlayerState.SUBTITLECHANGE, getPosition());
+                stateMachine.transitionState(originalState, getPosition());
+            }
+        }
+    };
+
     private OnStallStartedListener onStallStartedListener = new OnStallStartedListener() {
         @Override
         public void onStallStarted(StallStartedEvent stallStartedEvent) {
@@ -307,7 +353,8 @@ public class BitmovinSdkAdapter implements PlayerAdapter {
 
     private OnVideoPlaybackQualityChangedListener onVideoPlaybackQualityChangedListener = new OnVideoPlaybackQualityChangedListener() {
         @Override
-        public void onVideoPlaybackQualityChanged(VideoPlaybackQualityChangedEvent videoPlaybackQualityChangedEvent) {
+        public void onVideoPlaybackQualityChanged(
+            VideoPlaybackQualityChangedEvent videoPlaybackQualityChangedEvent) {
             Log.d(TAG, "On Video Quality Changed");
             if ((stateMachine.getCurrentState() == PlayerState.PLAYING) || (stateMachine.getCurrentState() == PlayerState.PAUSE)) {
                 PlayerState originalState = stateMachine.getCurrentState();
@@ -326,7 +373,8 @@ public class BitmovinSdkAdapter implements PlayerAdapter {
 
     private OnAudioPlaybackQualityChangedListener onAudioPlaybackQualityChangedListener = new OnAudioPlaybackQualityChangedListener() {
         @Override
-        public void onAudioPlaybackQualityChanged(AudioPlaybackQualityChangedEvent audioPlaybackQualityChangedEvent) {
+        public void onAudioPlaybackQualityChanged(
+            AudioPlaybackQualityChangedEvent audioPlaybackQualityChangedEvent) {
             Log.d(TAG, "On Audio Quality Changed");
             if ((stateMachine.getCurrentState() == PlayerState.PLAYING) || (stateMachine.getCurrentState() == PlayerState.PAUSE)) {
                 PlayerState originalState = stateMachine.getCurrentState();