+++ collector/src/main/java/com/bitmovin/analytics/data/LanguageInformation.kt
@@ -1,0 +1,8 @@
+package com.bitmovin.analytics.data
+
+
+data class LanguageInformation(var subtitleLanguage: String? = null, var audioLanguage: String? = null)
+{
+    val subtitleEnabled
+        get() = this.subtitleLanguage != null
+}