+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Development
 
+### Added
+- Sending out information about selected audio track language and subtitle language for the Bitmovin Player.
+
+
 ## v1.7.0
 
 ### Added